# =act improve ByteString#dropRight(...)