--- a/akka-bench-jmh/src/main/scala/akka/util/ByteString_dropRight_Benchmark.scala
+++ b/akka-bench-jmh/src/main/scala/akka/util/ByteString_dropRight_Benchmark.scala
@@ -49,32 +49,22 @@ class ByteString_dropRight_Benchmark {
    */
 
   @Benchmark
-  def bss_negative(): Unit = {
-    @volatile var bs: ByteString = null
-    bs = bss.dropRight(n_neg)
-  }
+  def bss_negative(): ByteString =
+    bss.dropRight(n_neg)
 
   @Benchmark
-  def bss_greater_or_eq_to_len(): Unit = {
-    @volatile var bs: ByteString = null
-    bs = bss.dropRight(n_greater_or_eq_to_len)
-  }
+  def bss_greater_or_eq_to_len(): ByteString =
+    bss.dropRight(n_greater_or_eq_to_len)
 
   @Benchmark
-  def bss_avg(): Unit = {
-    @volatile var bs: ByteString = null
-    bs = bss.dropRight(n_avg)
-  }
+  def bss_avg(): ByteString =
+    bss.dropRight(n_avg)
 
   @Benchmark
-  def bss_best(): Unit = {
-    @volatile var bs: ByteString = null
-    bs = bss.dropRight(n_best)
-  }
+  def bss_best(): ByteString =
+    bss.dropRight(n_best)
 
   @Benchmark
-  def bss_worst(): Unit = {
-    @volatile var bs: ByteString = null
-    bs = bss.dropRight(n_worst)
-  }
+  def bss_worst(): ByteString =
+    bss.dropRight(n_worst)
 }