Currently, we recursively call `dropRight(...)` and obtain the rest of `bytestrings` by calling `Vector#init`, which would not be quite efficient.

Instead of doing this, we can do as follows so that we can achieve better performance:
1. Seek the index of _last_ vector element we need to _dropRight_
2. Find the number of characters left to _dropRight_ from the _last_ ByteString1 element.
3. Create `ByteString` based on the information we obtained from 1 and 2

Then we just need to create a new `Vector[ByteString1]` at most twice, which should be better than the current implementation, i.e., create the rest `bytestrings` every time we check `bytestrings(Vector[ByteString1])` element, which ends up O(N) _init_ execution where _N_ is the length of `bytestrings`.
### Benchmark

I measured the performance using sbt-jmh with this [code](https://gist.github.com/monkey-mas/684ec6a5e2a0ce42f28417dcd53c9636).  Benchmark code was executed with some parameters such as `-t1 -f2 -wi 10 -i 20`.
#### Test environment

```
OS: Ubuntu 14.0.4 (Linux kernel version 3.13.0-95-generic)
CPU: Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz
VM version: JDK 1.8.0_77
```
#### Before

```
Benchmark                                  Mode  Cnt           Score         Error  Units
DropRight.dropRight_avg                   thrpt   40       26184.435 ±     403.058  ops/s
DropRight.dropRight_best                  thrpt   40     8707856.806 ±   50699.447  ops/s
DropRight.dropRight_greater_or_eq_to_len  thrpt   40       12289.742 ±     149.310  ops/s
DropRight.dropRight_negative              thrpt   40  1191630555.447 ± 5225134.083  ops/s
DropRight.dropRight_worst                 thrpt   40       12329.187 ±     283.525  ops/s
```
#### After

```
DropRight.dropRight_avg                   thrpt   40      630595.880 ±     3895.122  ops/s
DropRight.dropRight_best                  thrpt   40     7608841.379 ±   264211.274  ops/s
DropRight.dropRight_greater_or_eq_to_len  thrpt   40   877987542.956 ± 12697839.317  ops/s
DropRight.dropRight_negative              thrpt   40  1125036409.900 ± 17180858.452  ops/s
DropRight.dropRight_worst                 thrpt   40      224135.469 ±     4563.056  ops/s
```