@ktoso 2016-09-12
LGTM, thanks a lot for addressing this.
dropRight wasn't on hot-paths we cared about so indeed it was implemented a bit more leisurely, good to get it up to speed as well :)

Would you mind hiding the dropRight0 please?