@monkey-mas 2016-09-25
```
[error] (akka-typed-experimental/pr-validation:test) sbt.TestsFailedException: Tests unsuccessful
[error] Total time: 4721 s, completed Sep 25, 2016 4:36:27 PM
```
Umm, I think the error isn't related to my code but this happened twice in a row... :(